backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
pillow==10.4.0
